[SPARK-42364][PS][TESTS] Split 'pyspark.pandas.tests.test_dataframe'
### What changes were proposed in this pull request?
Split 'pyspark.pandas.tests.test_dataframe'

### Why are the changes needed?
'pyspark.pandas.tests.test_dataframe' is too large: more than 7K lines, may take >15min sometimes

### Does this PR introduce _any_ user-facing change?
No, test-only

### How was this patch tested?
updated UT

in my local env: 175 sec -> 95 sec + 102 sec

Closes #39915 from zhengruifeng/ps_split_test_df.

Authored-by: Ruifeng Zheng <ruifengz@apache.org>
